refactor(filters): Split out url filters
